Spring03 에 예외 Controller 예제 추가함

+++ Spring03_Final/src/main/java/com/gura/spring03/cafe/service/CafeServiceImpl.java
@@ -255,6 +255,14 @@ public class CafeServiceImpl implements CafeService{
 	@Override
 	public void deleteComment(HttpServletRequest request) {
 		int num=Integer.parseInt(request.getParameter("num"));
+		//삭제할 댓글 정보를 읽어와서 
+		CafeCommentDto dto=cafeCommentDao.getData(num);
+		String id=(String)request.getSession().getAttribute("id");
+		//글 작성자와 로그인된 아이디와 일치하지 않으면
+		if(!dto.getWriter().equals(id)) {
+			throw new NotDeleteException("남의 댓글 지우면 혼난당!");
+		}
+		
 		cafeCommentDao.delete(num);
 	}
